Add gprof.py to generate a gprof file from ITM traces
The DWT provide a PC sampler that can be used some
basic profiling informations but that usually not enough.
Gprof is can provide more informations but using it
but doing the profiling inside the microcontroller
use too munch memory.
Instead, we could complete DWT by adding software trace
that send the PC callee and caller and then post process
the traces to generate a gprof file.

Signed-off-by: Alexandre Bailon <abailon@baylibre.com>